+++ packages/core/src/index.ts
@@ -13,4 +13,4 @@ export * from './event/event.interface';
 export * from './server/server.injector';
 export * from './server/server.event';
 export { createServer, CreateServerConfig } from './server/server.factory';
-export { HttpListenerConfig, httpListener } from './http.listener';
+export { HttpListenerConfig, httpListener } from './listener/http.listener';
